+++ org.eclipse.draw2d/src/org/eclipse/draw2d/PrintFigureOperation.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2000, 2010 IBM Corporation and others.
+ * Copyright (c) 2000, 2025 IBM Corporation and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -122,7 +122,7 @@ public class PrintFigureOperation extends PrintOperation {
 	 */
 	@Override
 	protected void printPages() {
-		Graphics graphics = getFreshPrinterGraphics();
+		Graphics graphics = getFreshGraphics();
 		IFigure figure = getPrintSource();
 		setupPrinterGraphicsFor(graphics, figure);
 		Rectangle bounds = figure.getBounds();